General borrado y validacion de rol en el gnerealRoot
