--- FormGeneral.cs
@@ -1,69 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace chocoflantastico
-{
-    public partial class FormGeneral : Form
-    {
-        public FormGeneral(string nombre)
-        {
-            InitializeComponent();
-            lblUser.Text = nombre;
-        }
-
-        private void pbCerrarSesion_Click(object sender, EventArgs e)
-        {
-            this.Close();
-        }
-
-        private void btnCerrarSesion_Click(object sender, EventArgs e)
-        {
-            this.Close();
-        }
-
-        private void pbCerrar_Click(object sender, EventArgs e)
-        {
-            DialogResult result = MessageBox.Show("¿Está seguro que desea cerrar la aplicación", "CHOCOFLANTASTICO", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
-            if (result == DialogResult.OK)
-            {
-                Application.Exit();
-            }
-        }
-
-        private void btnCerrarSesion_Click_1(object sender, EventArgs e)
-        {
-            DialogResult result = MessageBox.Show("¿Está seguro que desea cerrar la aplicación", "CHOCOFLANTASTICO", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
-            if (result == DialogResult.OK)
-            {
-                Application.Exit();
-            }
-
-        }
-
-        private void pbCerrarSesion_Click_1(object sender, EventArgs e)
-        {
-            DialogResult result = MessageBox.Show("¿Está seguro que desea cerrar la aplicación", "CHOCOFLANTASTICO", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
-            if (result == DialogResult.OK)
-            {
-                Application.Exit();
-            }
-        }
-
-        private void pbMinimizar_Click(object sender, EventArgs e)
-        {
-            this.WindowState = FormWindowState.Minimized;
-        }
-
-        private void pnlForm_Paint(object sender, PaintEventArgs e)
-        {
-
-        }
-    }
-}
