+++ FormUsers.cs
@@ -15,6 +15,29 @@ namespace chocoflantastico
         public FormUsers()
         {
             InitializeComponent();
+            if (dgvUsers.SelectedRows.Count == 0)
+            {
+                btnInhabilitar.Enabled = false;
+                btnModificar.Enabled = false;
+            }
+        }
+
+        private void btnCerrarSesion_Click(object sender, EventArgs e)
+        {
+            DialogResult result = MessageBox.Show("¿Está seguro que desea cerrar la aplicación", "CHOCOFLANTASTICO", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+            if (result == DialogResult.OK)
+            {
+                Application.Exit();
+            }
+        }
+
+        private void pbCerrar_Click(object sender, EventArgs e)
+        {
+            DialogResult result = MessageBox.Show("¿Está seguro que desea cerrar la aplicación", "CHOCOFLANTASTICO", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
+            if (result == DialogResult.OK)
+            {
+                Application.Exit();
+            }
         }
     }
 }
